restructure endpoints
